Removed smelly code components. Code is now much more abstract.
